refactor: create a separate component for the list item
